+++ src/main/java/mdteam/ait/core/item/SonicItem.java
@@ -115,6 +115,9 @@ public class SonicItem extends LinkableItem {
                     if (!(exteriorBlockEntity.findTardis().get() == tardis)) return;
                     int repairticksleft = tardis.getHandlers().getCrashData().getRepairTicks();
                     int repairminutes = repairticksleft / 20 / 60;
+                    if (repairticksleft == 0) {
+                        player.sendMessage(Text.literal("You tardis is not damaged"));
+                    }
                     player.sendMessage(Text.literal("You have " + repairminutes + " minutes of repair left.").formatted(Formatting.GOLD), true);
                     return;
                 }