+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 # 1.2.42
+- fixing tardis flight and crash data