+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.2.42
 - fixing tardis flight and crash data
-- sonic can see repair time left
+- sonic can see repair time left, shift right click on the exterior